Switch to SDMemoryCache
